requirements.txt
numpy==2.1.1
openai==1.50.2
openai_whisper==20231117
PyAutoGUI==0.9.54
python-dotenv==1.0.1
pyttsx3==2.97
sounddevice==0.5.0
spotipy==2.24.0
webrtcvad==2.0.10
